requirements.txt
fastapi
uvicorn
streamlit

crewai
langchain
langchain-community

chromadb

python-dotenv

tavily-python

groq
pydantic
requests